# ARROW-10660: [Rust] Implement AVX-512 bit or operation