Implements AVX-512 bit or operation

Before
```
buffer_bit_ops or       time:   [681.78 ns 682.42 ns 683.36 ns]                               
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) low severe
  4 (4.00%) low mild
  2 (2.00%) high mild
  7 (7.00%) high severe
```

After
```
buffer_bit_ops or       time:   [536.78 ns 537.75 ns 539.16 ns]                               
                        change: [-21.354% -20.879% -20.123%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
```